+++ service/src/main/java/com/senla/socialnetwork/service/PublicMessageServiceImpl.java
@@ -1,11 +1,13 @@
 package com.senla.socialnetwork.service;
 
 import com.senla.socialnetwork.dao.LocationDao;
+import com.senla.socialnetwork.dao.PublicMessageCommentDao;
 import com.senla.socialnetwork.dao.PublicMessageDao;
 import com.senla.socialnetwork.dao.SchoolDao;
 import com.senla.socialnetwork.dao.UniversityDao;
 import com.senla.socialnetwork.dao.UserProfileDao;
 import com.senla.socialnetwork.domain.PublicMessage;
+import com.senla.socialnetwork.domain.PublicMessageComment;
 import com.senla.socialnetwork.dto.PublicMessageDto;
 import com.senla.socialnetwork.service.exception.BusinessException;
 import com.senla.socialnetwork.service.util.PublicMessageMapper;
@@ -21,8 +23,12 @@ import java.util.List;
 @NoArgsConstructor
 @Slf4j
 public class PublicMessageServiceImpl implements PublicMessageService {
+    private static final int FIRST_RESULT = 0;
+    private static final int MAX_RESULTS = 0;
     @Autowired
     PublicMessageDao publicMessageDao;
+    @Autowired
+    PublicMessageCommentDao publicMessageCommentDao;
     @Autowired
     UserProfileDao userProfileDao;
     @Autowired
@@ -31,6 +37,14 @@ public class PublicMessageServiceImpl implements PublicMessageService {
     SchoolDao schoolDao;
     @Autowired
     UniversityDao universityDao;
+
+    @Override
+    @Transactional
+    public List<PublicMessageDto> getMessages(int firstResult, int maxResults) {
+        log.debug("[getSchools]");
+        log.debug("[firstResult: {}, maxResults: {}]", firstResult, maxResults);
+        return PublicMessageMapper.getPublicMessageDto(publicMessageDao.getAllRecords(firstResult, maxResults));
+    }
 
     @Override
     @Transactional
@@ -76,7 +90,6 @@ public class PublicMessageServiceImpl implements PublicMessageService {
 
     @Override
     @Transactional
-    // ToDo check comments, need to delete it?
     public void deleteMessageByUser(String email, Long messageId) {
         log.debug("[deleteMessageByUser]");
         log.debug("[email: {}, messageId: {}]", email, messageId);
@@ -86,6 +99,10 @@ public class PublicMessageServiceImpl implements PublicMessageService {
         } else if (publicMessage.isDeleted()) {
             throw new BusinessException("Error, the message has already been deleted");
         }
+        List<PublicMessageComment> comments = publicMessageCommentDao.getPublicMessageComments(
+            messageId, FIRST_RESULT, MAX_RESULTS);
+        comments.forEach(publicMessageComment -> publicMessageComment.setDeleted(true));
+        publicMessage.setPublicMessageComments(comments);
         publicMessage.setDeleted(true);
         publicMessageDao.updateRecord(publicMessage);
     }
@@ -95,6 +112,9 @@ public class PublicMessageServiceImpl implements PublicMessageService {
     public void deleteMessage(Long messageId) {
         log.debug("[deleteMessage]");
         log.debug("[messageId: {}]", messageId);
+        if (publicMessageDao.findById(messageId) == null) {
+            throw new BusinessException("Error, there is no such message");
+        }
         publicMessageDao.deleteRecord(messageId);
     }
